Remove wiki

+++ db/migrate/20240508163803_drop_unneeded_tables.rb
@@ -1,8 +1,8 @@
 class DropUnneededTables < ActiveRecord::Migration[7.1]
   def up
-    drop_table :comments
-    drop_table :page_versions
-    drop_table :pages
+    drop_table :comments if table_exists? :comments
+    drop_table :page_versions if table_exists? :page_versions
+    drop_table :pages if table_exists? :pages
   end
 
   def down
